move extracted components into page components
