--- packages/builder/src/components/login/index.js
@@ -1,1 +1,0 @@
-export { LoginForm } from "./LoginForm.svelte"
