+++ packages/builder/src/pages/builder/auth/login.svelte
@@ -1,5 +1,84 @@
 <script>
-  import LoginForm from "components/login/LoginForm.svelte"
+  import {
+    ActionButton,
+    Body,
+    Button,
+    Divider,
+    Heading,
+    Input,
+    Layout,
+    notifications,
+  } from "@budibase/bbui"
+  import { goto } from "@roxi/routify"
+  import { auth, organisation } from "stores/portal"
+  import GoogleButton from "./_components/GoogleButton.svelte"
+
+  let username = ""
+  let password = ""
+
+  async function login() {
+    try {
+      await auth.login({
+        username,
+        password,
+      })
+      notifications.success("Logged in successfully")
+      if ($auth?.user?.forceResetPassword) {
+        $goto("./reset")
+      } else {
+        $goto("../portal")
+      }
+    } catch (err) {
+      console.error(err)
+      notifications.error("Invalid credentials")
+    }
+  }
 </script>
 
-<LoginForm />
+<div class="login">
+  <div class="main">
+    <Layout>
+      <Layout noPadding justifyItems="center">
+        <img src={$organisation.logoUrl || "https://i.imgur.com/ZKyklgF.png"} />
+        <Heading>Sign in to Budibase</Heading>
+      </Layout>
+      <GoogleButton />
+      <Divider noGrid />
+      <Layout gap="XS" noPadding>
+        <Body size="S" textAlign="center">Sign in with email</Body>
+        <Input label="Email" bind:value={username} />
+        <Input
+          label="Password"
+          type="password"
+          on:change
+          bind:value={password}
+        />
+      </Layout>
+      <Layout gap="XS" noPadding>
+        <Button cta on:click={login}>Sign in to Budibase</Button>
+        <ActionButton quiet on:click={() => $goto("./forgot")}>
+          Forgot password?
+        </ActionButton>
+      </Layout>
+    </Layout>
+  </div>
+</div>
+
+<style>
+  .login {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .main {
+    width: 300px;
+  }
+
+  img {
+    width: 48px;
+  }
+</style>
